HostWallet, AccountSigning and EIP712 Encoding in ABI
+ New host wallet rpc methods (AddChain, RequestPermissions, GetPermissions
+ New AccountSigning rpc methods for persionalSign and TypedData V4, this will get replace when using an Account or External Account to sign with Nethereum or other External Account provider (ie AzureKeyVault, Trezoer, Ledger)
+ EIP712 Encoding is part of the the Abi library
+ Metamask interceptor changes for v4 and personalSign to add from account

+++ src/Nethereum.Metamask.Blazor/MetamaskBlazorInterop.cs
@@ -71,7 +71,7 @@ namespace Nethereum.Metamask.Blazor
         }
 
         [JSInvokable()]
-        public static async Task SelectedNetworkChanged(int chainId)
+        public static async Task SelectedNetworkChanged(long chainId)
         {
             await MetamaskHostProvider.Current.ChangeSelectedNetworkAsync(chainId).ConfigureAwait(false);
         }
